Add template projects and fix example compilation errors

+++ examples/BasicInput/BasicInputSimulation.cs
@@ -13,7 +13,7 @@ internal class BasicInputSimulation : Simulation
     {
     }
 
-    int c;
+    float c;
     Vector2 p;
 
     public override void OnRender(ICanvas canvas)
